add compareRevisions test
